+++ packages/framework-core/src/booster-event-dispatcher.ts
@@ -12,11 +12,12 @@ import { ReadModelStore } from './services/read-model-store'
 import { RegisterHandler } from './booster-register-handler'
 import { BoosterGlobalErrorDispatcher } from './booster-global-error-dispatcher'
 import { createInstance, Promises, getLogger } from '@boostercloud/framework-common-helpers'
+import { NotificationInterface } from 'framework-types/dist'
 
 export class BoosterEventDispatcher {
   /**
    * Entry point to dispatch events coming from the cloud provider.
-   * @param rawEvents List of RAW events from the cloud provider
+   * @param rawEvents List of raw events from the cloud provider
    * @param config
    * @param logger
    */
@@ -42,16 +43,18 @@ export class BoosterEventDispatcher {
    */
   private static eventProcessor(eventStore: EventStore, readModelStore: ReadModelStore): EventsStreamingCallback {
     return async (entityName, entityID, eventEnvelopes, config) => {
-      await Promises.allSettledAndFulfilled([
-        await BoosterEventDispatcher.snapshotAndUpdateReadModels(
-          config,
-          entityName,
-          entityID,
-          eventStore,
-          readModelStore
-        ),
-        await BoosterEventDispatcher.dispatchEntityEventsToEventHandlers(eventEnvelopes, config),
-      ])
+      const eventEnvelopesProcessors = [
+        BoosterEventDispatcher.dispatchEntityEventsToEventHandlers(eventEnvelopes, config),
+      ]
+
+      // Read models are not updated for notification events (events that are not related to an entity but a topic)
+      if (!(entityName in config.topicToEvent)) {
+        eventEnvelopesProcessors.push(
+          BoosterEventDispatcher.snapshotAndUpdateReadModels(config, entityName, entityID, eventStore, readModelStore)
+        )
+      }
+
+      await Promises.allSettledAndFulfilled(eventEnvelopesProcessors)
     }
   }
 
@@ -78,7 +81,7 @@ export class BoosterEventDispatcher {
   }
 
   private static async dispatchEntityEventsToEventHandlers(
-    entityEventEnvelopes: Array<EventEnvelope>,
+    entityEventEnvelopes: Array<EventEnvelope | NotificationInterface>,
     config: BoosterConfig
   ): Promise<void> {
     const logger = getLogger(config, 'BoosterEventDispatcher.dispatchEntityEventsToEventHandlers')
@@ -88,7 +91,7 @@ export class BoosterEventDispatcher {
         logger.debug(`No event-handlers found for event ${eventEnvelope.typeName}. Skipping...`)
         continue
       }
-      const eventClass = config.events[eventEnvelope.typeName]
+      const eventClass = config.events[eventEnvelope.typeName] ?? config.notifications[eventEnvelope.typeName]
       await Promises.allSettledAndFulfilled(
         eventHandlers.map(async (eventHandler: EventHandlerInterface) => {
           const eventInstance = createInstance(eventClass.class, eventEnvelope.value)