+++ packages/framework-core/test/booster-event-dispatcher.test.ts
@@ -30,12 +30,17 @@ class SomeEvent {
   }
 }
 
+class SomeNotification {
+  public constructor() {}
+}
+
 class AnEventHandler {
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   public static async handle(event: SomeEvent, register: Register): Promise<void> {
     event.getPrefixedId('prefix')
   }
 }
+
 const someEvent: NonPersistedEventEnvelope = {
   version: 1,
   kind: 'event',
@@ -49,7 +54,18 @@ const someEvent: NonPersistedEventEnvelope = {
     id: '42',
   },
   requestID: '123',
-  typeName: 'SomeEvent',
+  typeName: SomeEvent.name,
+}
+
+const someNotification: NonPersistedEventEnvelope = {
+  version: 1,
+  kind: 'event',
+  superKind: 'domain',
+  entityID: 'default',
+  entityTypeName: 'defaultTopic',
+  value: {},
+  requestID: '123',
+  typeName: SomeNotification.name,
 }
 
 const someEntity: EntityInterface = {
@@ -78,6 +94,7 @@ describe('BoosterEventDispatcher', () => {
   const config = new BoosterConfig('test')
   config.provider = {} as ProviderLibrary
   config.events[SomeEvent.name] = { class: SomeEvent }
+  config.notifications[SomeNotification.name] = { class: SomeNotification }
   config.logger = {
     info: fake(),
     error: fake(),
@@ -153,6 +170,25 @@ describe('BoosterEventDispatcher', () => {
           [someEvent],
           config
         )
+      })
+
+      it("doesn't call snapshotAndUpdateReadModels if the entity name is in config.topicToEvent", async () => {
+        const stubEventStore = createStubInstance(EventStore)
+        const stubReadModelStore = createStubInstance(ReadModelStore)
+
+        const boosterEventDispatcher = BoosterEventDispatcher as any
+        replace(boosterEventDispatcher, 'snapshotAndUpdateReadModels', fake())
+        replace(boosterEventDispatcher, 'dispatchEntityEventsToEventHandlers', fake())
+
+        const overriddenConfig = { ...config }
+        overriddenConfig.topicToEvent = { [someEvent.entityTypeName]: 'SomeEvent' }
+
+        const callback = boosterEventDispatcher.eventProcessor(stubEventStore, stubReadModelStore)
+
+        await callback(someEvent.entityTypeName, someEvent.entityID, [someEvent], overriddenConfig)
+        overriddenConfig.topicToEvent = {}
+
+        expect(boosterEventDispatcher.snapshotAndUpdateReadModels).not.to.have.been.called
       })
     })
 
@@ -254,6 +290,22 @@ describe('BoosterEventDispatcher', () => {
 
         expect(fakeHandler1).to.have.been.calledOnceWith(anEventInstance)
         expect(fakeHandler2).to.have.been.calledOnceWith(anEventInstance)
+      })
+
+      it('calls all the handlers, even if the event is stored in the notifications field instead of the events one', async () => {
+        const fakeHandler1 = fake()
+        const fakeHandler2 = fake()
+        config.eventHandlers[SomeNotification.name] = [{ handle: fakeHandler1 }, { handle: fakeHandler2 }]
+
+        replace(RegisterHandler, 'handle', fake())
+
+        const boosterEventDispatcher = BoosterEventDispatcher as any
+        await boosterEventDispatcher.dispatchEntityEventsToEventHandlers([someNotification], config)
+
+        const aNotificationInstance = new SomeNotification()
+
+        expect(fakeHandler1).to.have.been.calledOnceWith(aNotificationInstance)
+        expect(fakeHandler2).to.have.been.calledOnceWith(aNotificationInstance)
       })
 
       it('calls the register handler for all the published events', async () => {