+++ packages/framework-core/test/booster-register-handler.test.ts
@@ -17,6 +17,10 @@ class SomeEvent {
   }
 }
 
+class SomeNotification {
+  public constructor() {}
+}
+
 describe('the `RegisterHandler` class', () => {
   const testConfig = new BoosterConfig('Test')
   testConfig.logLevel = Level.debug
@@ -141,6 +145,35 @@ describe('the `RegisterHandler` class', () => {
       value: event,
       currentUser: user,
       typeName: 'SomeEvent',
+    })
+  })
+
+  it('can wrap notifications to produce eventEnvelopes', () => {
+    const config = new BoosterConfig('test')
+    config.notifications[SomeNotification.name] = {
+      class: SomeNotification,
+    }
+
+    const user: UserEnvelope = {
+      username: 'paco@example.com',
+      roles: ['Paco'],
+      claims: {},
+    }
+
+    const register = new Register('1234', {} as any, RegisterHandler.flush, user)
+    const notification = new SomeNotification()
+
+    const registerHandler = RegisterHandler as any
+    expect(registerHandler.wrapEvent(config, notification, register)).to.deep.equal({
+      version: 1,
+      kind: 'event',
+      superKind: 'domain',
+      entityID: 'default',
+      requestID: '1234',
+      entityTypeName: 'defaultTopic',
+      value: notification,
+      currentUser: user,
+      typeName: SomeNotification.name,
     })
   })