+++ packages/framework-types/src/envelope.ts
@@ -1,4 +1,12 @@
-import { CommandInput, EntityInterface, EventInterface, ReadModelInterface, SequenceKey, UUID } from './concepts'
+import {
+  CommandInput,
+  EntityInterface,
+  EventInterface,
+  NotificationInterface,
+  ReadModelInterface,
+  SequenceKey,
+  UUID,
+} from './concepts'
 import { GraphQLClientMessage } from './graphql-websocket-messages'
 import { FilterFor, SortFor } from './searcher'
 import { Class } from './typelevel'
@@ -32,12 +40,13 @@ export interface EventStoreEntryEnvelope extends Envelope {
   superKind: SuperKindType
   entityID: UUID
   entityTypeName: string
-  value: EventInterface | EntityInterface
+  value: EventInterface | EntityInterface | NotificationInterface
 }
 
 export interface NonPersistedEventEnvelope extends EventStoreEntryEnvelope {
   kind: 'event'
 }
+
 export interface EventEnvelope extends NonPersistedEventEnvelope {
   createdAt: string
 }